requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.7.5
aniso8601==9.0.1
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.3
attrs==23.1.0
backports.entry-points-selectable==1.1.1
black==21.12b0
blinker==1.6.2
boto3==1.20.28
botocore==1.23.51
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.0.9
click==8.1.7
contourpy==1.0.6
customtkinter==5.0.3
cycler==0.11.0
darkdetect==0.7.1
distlib==0.3.4
Django==4.0.1
fastapi==0.70.1
filelock==3.4.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
fonttools==4.38.0
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.12.0
identify==2.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
kiwisolver==1.4.4
Mako==1.1.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.6.2
multidict==6.0.4
mypy-extensions==0.4.3
mysqlclient==2.1.0
nodeenv==1.6.0
numpy==1.23.5
openai==0.27.8
packaging==22.0
pathspec==0.9.0
Pillow==9.3.0
pip-chill==1.0.3
pipdeptree==2.13.0
platformdirs==2.4.0
pre-commit==2.16.0
pydantic==1.8.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
s3transfer==0.5.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
sqlparse==0.4.2
starlette==0.16.0
toml==0.10.2
tomli==1.2.3
tqdm==4.66.1
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
virtualenv==20.10.0
Werkzeug==2.3.7
yarl==1.9.2
